searches db for links
The reverse of dbfiles. Drop this into a folder eg. files or images and
it will search through the db and report on the links found for that
folder and also whether the file is found in that folder (or subfolder).
Gives a table showing the link, the db tables, id, published state and
if the file is found a link to the file.